+++ src/ovito/stdobj/io/DataTableExporter.cpp
@@ -145,7 +145,7 @@ bool DataTableExporter::exportFrame(int frameNumber, const QString& filePath, Ma
             if(xaccessInt8)
                 textStream() << static_cast<qint32>(xaccessInt8.get(row, 0)) << " ";
             else if(xaccessInt32)
-                textStream() << xaccessInt32.get(row, 0) << " ";
+                textStream() << static_cast<qint32>(xaccessInt32.get(row, 0)) << " ";
             else if(xaccessInt64)
                 textStream() << static_cast<qint64>(xaccessInt64.get(row, 0)) << " ";
             else if(xaccessFloat32)
@@ -161,7 +161,7 @@ bool DataTableExporter::exportFrame(int frameNumber, const QString& filePath, Ma
                 if(array.dataType() == PropertyObject::Int8)
                     textStream() << static_cast<qint32>(*reinterpret_cast<const int8_t*>(array.cdata(row, col))) << " ";
                 else if(array.dataType() == PropertyObject::Int32)
-                    textStream() << *reinterpret_cast<const int32_t*>(array.cdata(row, col)) << " ";
+                    textStream() << static_cast<qint32>(*reinterpret_cast<const int32_t*>(array.cdata(row, col))) << " ";
                 else if(array.dataType() == PropertyObject::Int64)
                     textStream() << static_cast<qint64>(*reinterpret_cast<const int64_t*>(array.cdata(row, col))) << " ";
                 else if(array.dataType() == PropertyObject::Float32)
